Use template module in pltr to apply template on file creation
